Refactor web handling and remove obsolete page templates
- Updated the NewEcho function to utilize a new PageHandler for route registration, enhancing modularity.
- Removed the old PageHandler implementation and associated page templates (home.templ, contact.templ, base.templ) to streamline the codebase.
- Introduced a new Renderer for handling template rendering, improving separation of concerns.
- Updated web module to include the new view module, ensuring proper dependency management.

These changes improve the maintainability and organization of the web handling components, aligning with modern development practices.
