--- internal/layouts/base.templ
@@ -1,35 +1,0 @@
-// internal/layouts/base.templ
-package layouts
-
-templ Base(title string) {
-    <!DOCTYPE html>
-    <html lang="en">
-        <head>
-            <meta charset="UTF-8"/>
-            <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-            <meta name="description" content="Goforms - A self-hosted form backend service built with Go"/>
-            <meta name="keywords" content="forms, golang, backend, self-hosted"/>
-            <title>{ title }</title>
-            <link rel="stylesheet" href="/static/css/main.css"/>
-        </head>
-        <body>
-            <nav>
-                <div class="container">
-                    <a href="/" class="logo">Goforms</a>
-                    <div class="nav-links">
-                        <a href="/contact">Contact Demo</a>
-                        <a href="https://github.com/jonesrussell/goforms">GitHub</a>
-                    </div>
-                </div>
-            </nav>
-
-            { children... }
-
-            <footer>
-                <div class="container">
-                    Created by <a href="https://jonesrussell.github.io/me/" target="_blank" rel="noopener noreferrer">Russell Jones</a>
-                </div>
-            </footer>
-        </body>
-    </html>
-}
