--- internal/pages/contact.templ
@@ -1,20 +1,0 @@
-// internal/pages/contact.templ
-
-package pages
-
-import (
-    "github.com/jonesrussell/goforms/internal/layouts"
-    "github.com/jonesrussell/goforms/internal/components"
-)
-
-templ Contact() {
-    @layouts.Base("Contact Demo - Goforms") {
-        <main>
-            <section class="contact-section">
-                <div class="container">
-                    @components.ContactForm()
-                </div>
-            </section>
-        </main>
-    }
-}
