--- internal/pages/home.templ
@@ -1,35 +1,0 @@
-package pages
-
-import "github.com/jonesrussell/goforms/internal/layouts"
-
-templ Home() {
-    @layouts.Base("Goforms - Form Backend") {
-        <main>
-            <section class="hero">
-                <div class="container">
-                    <h1>Free Form Backend Service</h1>
-                    <p>Use our free hosted service or deploy your own. No dependencies, just Go.</p>
-                    <div class="cta-buttons">
-                        <a href="/contact" class="button primary">Try Demo Form</a>
-                        <a href="https://github.com/jonesrussell/goforms" class="button">View on GitHub</a>
-                    </div>
-                </div>
-            </section>
-
-            <section class="features container">
-                <div class="feature-card">
-                    <h3>🎯 Free Hosted</h3>
-                    <p>Start using our service instantly, no setup required.</p>
-                </div>
-                <div class="feature-card">
-                    <h3>🔒 Self Hosted Option</h3>
-                    <p>Deploy on your own infrastructure if needed.</p>
-                </div>
-                <div class="feature-card">
-                    <h3>📡 API First</h3>
-                    <p>RESTful API with JSON responses.</p>
-                </div>
-            </section>
-        </main>
-    }
-}
