+++ internal/view/module.go
@@ -1,15 +1,12 @@
-package web
+package view
 
 import (
 	"go.uber.org/fx"
-
-	"github.com/jonesrussell/goforms/internal/view"
 )
 
 //nolint:gochecknoglobals // This is an intentional global following fx module pattern
 var Module = fx.Options(
 	fx.Provide(
-		NewPageHandler,
+		NewRenderer,
 	),
-	view.Module,
 )
