+++ internal/web/handler.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/labstack/echo/v4"
 
+	"github.com/jonesrussell/goforms/internal/components"
 	"github.com/jonesrussell/goforms/internal/logger"
-	"github.com/jonesrussell/goforms/internal/pages"
 	"github.com/jonesrussell/goforms/internal/view"
 )
 
@@ -27,9 +27,22 @@ func (h *PageHandler) wrapError(err error, msg string) error {
 }
 
 func (h *PageHandler) Home(c echo.Context) error {
-	if err := h.renderer.Render(c, pages.Home()); err != nil {
+	if err := h.renderer.Render(c, components.Home()); err != nil {
 		h.logger.Error("failed to render home page", logger.Error(err))
 		return h.wrapError(err, "failed to render home page")
 	}
 	return nil
+}
+
+func (h *PageHandler) Contact(c echo.Context) error {
+	if err := h.renderer.Render(c, components.Contact()); err != nil {
+		h.logger.Error("failed to render contact page", logger.Error(err))
+		return h.wrapError(err, "failed to render contact page")
+	}
+	return nil
+}
+
+func (h *PageHandler) RegisterRoutes(e *echo.Echo) {
+	e.GET("/", h.Home)
+	e.GET("/contact", h.Contact)
 }
